feat: 정보탭 지수 섹션 및 지도 오버레이 책임 분리
