+++ src/components/Info/section/StationIndexSection.tsx
@@ -1,0 +1,20 @@
+import StationIndexBar from "@/assets/icons/station_index_tip.svg?react";
+
+type StationIndexSectionProps = {
+  title: string;
+};
+
+const StationIndexSection = ({ title }: StationIndexSectionProps) => {
+  return (
+    <div className="flex flex-col pt-[15px] pl-5">
+      <div className="text-base font-semibold text-coolNeutral-10">
+        {title}
+      </div>
+      <div className="pt-[55px] pb-[83px]">
+        <StationIndexBar />
+      </div>
+    </div>
+  );
+};
+
+export default StationIndexSection;
