+++ src/components/Map.tsx
@@ -4,8 +4,11 @@ import seoulGeoJson from "@/assets/geojson/seoul-gu-simple.json";
 import { regionMap } from "@/constants/regionMap";
 import { regionColor, regionBoundaryColor } from "@/constants/regionColor";
 import { useMapOverlayStore } from "@/store/mapOverlayStore";
+import {
+  clearConsumerIndexPolygons,
+  drawConsumerIndexPolygons,
+} from "@/components/mapOverlays/consumerIndexOverlay";
 import type {
-  ConsumerIndexMapOverlayItem,
   RentIndexMapOverlayItem,
   RentIndexMapOverlayType,
 } from "@/store/mapOverlayStore";
@@ -30,18 +33,6 @@ const getOverlayColor = (type: RentIndexMapOverlayType) => {
   if (type === "RISE") return "#FF5555";
   if (type === "FALL") return "#4D8DFF";
   return "#3385FF";
-};
-
-const getConsumerIndexPhase = (value: number) => {
-  if (value <= 94) return "하강 국면";
-  if (value <= 114) return "보합 국면";
-  return "상승 국면";
-};
-
-const getConsumerIndexOverlayColor = (value: number) => {
-  if (value <= 94) return "#4D8DFF";
-  if (value <= 114) return "#F2C94C";
-  return "#FF5555";
 };
 
 const getOverlayValueLabel = (item: RentIndexMapOverlayItem) => {
@@ -313,13 +304,6 @@ const Map = ({ enableOverlay = true }: Props) => {
     selectedRegionPolygonsRef.current = [];
   };
 
-  const clearConsumerIndexPolygons = () => {
-    consumerIndexPolygonsRef.current.forEach((polygon) => {
-      polygon.setMap(null);
-    });
-    consumerIndexPolygonsRef.current = [];
-  };
-
   const drawSelectedRegionPolygons = (
     map: any,
     regionName: string,
@@ -431,83 +415,6 @@ const Map = ({ enableOverlay = true }: Props) => {
     map.panTo(new window.kakao.maps.LatLng(center.lat, center.lng));
   };
 
-  const drawConsumerIndexPolygons = (
-    map: any,
-    item: ConsumerIndexMapOverlayItem | null,
-  ) => {
-    clearConsumerIndexPolygons();
-
-    if (!item) return;
-
-    const fillColor = getConsumerIndexOverlayColor(item.value);
-    const phase = getConsumerIndexPhase(item.value);
-
-    seoulGeoJson.features.forEach((feature: any) => {
-      const { geometry, properties } = feature;
-      const guName = properties.SIG_KOR_NM;
-      const polygons =
-        geometry.type === "Polygon" ? [geometry.coordinates] : geometry.coordinates;
-
-      polygons.forEach((rings: number[][][]) => {
-        const polygon = new window.kakao.maps.Polygon({
-          map,
-          path: createPolygonPaths(rings),
-          strokeWeight: 2,
-          strokeColor: fillColor,
-          strokeOpacity: 0.95,
-          fillColor,
-          fillOpacity: 0.34,
-          zIndex: 3,
-        });
-
-        window.kakao.maps.event.addListener(
-          polygon,
-          "mouseover",
-          (mouseEvent: any) => {
-            polygon.setOptions({
-              fillOpacity: 0.52,
-            });
-
-            if (!overlayRef.current || !enableOverlay) return;
-
-            overlayRef.current.setContent(`
-              <div style="
-                padding:8px 18px;
-                background:white;
-                text-align:center;
-                border-radius:4px;
-                font-size:13px;
-                line-height:1.45;
-                font-family:'Pretendard', sans-serif;
-                box-shadow:0 2px 6px rgba(0,0,0,0.2);
-              ">
-                <strong>${guName}</strong><br/>
-                서울 소비자 심리지수<br/>
-                ${item.year}년 ${item.month}월 ${item.value.toFixed(1)}점<br/>
-                <span style="color:${fillColor}; font-weight:700;">${phase}</span>
-              </div>
-            `);
-
-            overlayRef.current.setPosition(mouseEvent.latLng);
-            overlayRef.current.setMap(map);
-          },
-        );
-
-        window.kakao.maps.event.addListener(polygon, "mouseout", () => {
-          polygon.setOptions({
-            fillOpacity: 0.34,
-          });
-
-          if (overlayRef.current) {
-            overlayRef.current.setMap(null);
-          }
-        });
-
-        consumerIndexPolygonsRef.current.push(polygon);
-      });
-    });
-  };
-
   // const createPolygon = (map: any, rings: number[][][], properties: any) => {
   //   const paths = rings.map((ring) =>
   //     ring.map(([lng, lat]) => new window.kakao.maps.LatLng(lat, lng)),
@@ -599,12 +506,19 @@ const Map = ({ enableOverlay = true }: Props) => {
   useEffect(() => {
     if (!isMapReady || !mapRefInstance.current) return;
 
-    drawConsumerIndexPolygons(mapRefInstance.current, consumerIndexItem);
+    drawConsumerIndexPolygons({
+      map: mapRefInstance.current,
+      item: consumerIndexItem,
+      overlay: overlayRef.current,
+      enableOverlay,
+      createPolygonPaths,
+      polygonRefs: consumerIndexPolygonsRef.current,
+    });
 
     return () => {
-      clearConsumerIndexPolygons();
+      clearConsumerIndexPolygons(consumerIndexPolygonsRef.current);
     };
-  }, [consumerIndexItem, isMapReady]);
+  }, [consumerIndexItem, enableOverlay, isMapReady]);
 
   useEffect(() => {
     if (!isMapReady || !mapRefInstance.current || !selectedRentIndexItem) {
